+++ server/src/main/java/org/opensearch/cluster/decommission/DecommissionService.java
@@ -13,11 +13,11 @@ import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.opensearch.OpenSearchTimeoutException;
 import org.opensearch.action.ActionListener;
+import org.opensearch.action.admin.cluster.decommission.awareness.put.DecommissionResponse;
 import org.opensearch.cluster.ClusterState;
 import org.opensearch.cluster.ClusterStateObserver;
 import org.opensearch.cluster.ClusterStateUpdateTask;
 import org.opensearch.cluster.NotClusterManagerException;
-import org.opensearch.cluster.ack.ClusterStateUpdateResponse;
 import org.opensearch.cluster.metadata.Metadata;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.routing.allocation.AllocationService;
@@ -117,7 +117,7 @@ public class DecommissionService {
      */
     public void startDecommissionAction(
         final DecommissionAttribute decommissionAttribute,
-        final ActionListener<ClusterStateUpdateResponse> listener
+        final ActionListener<DecommissionResponse> listener
     ) {
         // register the metadata with status as INIT as first step
         clusterService.submitStateUpdateTask("decommission [" + decommissionAttribute + "]", new ClusterStateUpdateTask(Priority.URGENT) {
@@ -163,7 +163,7 @@ public class DecommissionService {
 
     private synchronized void decommissionClusterManagerNodes(
         final DecommissionAttribute decommissionAttribute,
-        ActionListener<ClusterStateUpdateResponse> listener
+        ActionListener<DecommissionResponse> listener
     ) {
         ClusterState state = clusterService.getClusterApplierService().state();
         // since here metadata is already registered with INIT, we can guarantee that no new node with decommission attribute can further
@@ -183,7 +183,8 @@ public class DecommissionService {
         final Predicate<ClusterState> allNodesRemovedAndAbdicated = clusterState -> {
             final Set<String> votingConfigNodeIds = clusterState.getLastCommittedConfiguration().getNodeIds();
             return nodeIdsToBeExcluded.stream().noneMatch(votingConfigNodeIds::contains)
-                && nodeIdsToBeExcluded.contains(clusterState.nodes().getClusterManagerNodeId()) == false;
+                && nodeIdsToBeExcluded.contains(clusterState.nodes().getClusterManagerNodeId()) == false
+                && clusterState.nodes().getClusterManagerNodeId() != null;
         };
 
         ActionListener<Void> exclusionListener = new ActionListener<Void>() {
@@ -205,7 +206,7 @@ public class DecommissionService {
                         // we are good here to send the response now as the request is processed by an eligible active leader
                         // and to-be-decommissioned cluster manager is no more part of Voting Configuration and no more to-be-decommission
                         // nodes can be part of Voting Config
-                        listener.onResponse(new ClusterStateUpdateResponse(true));
+                        listener.onResponse(new DecommissionResponse(true));
                         failDecommissionedNodes(clusterService.getClusterApplierService().state());
                     }
                 } else {